Move BlueTestSummary cache into RunImpl
